requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
psycopg2-binary==2.9.3
django_filter
djangorestframework
